Mejoras de datos entre vista

+++ app/src/main/java/com/example/market/ui/products/model/ProductViewModel.kt
@@ -10,5 +10,5 @@ data class ProductViewModel(
     val unit: String = "VACIO",
     val price: Int = 0,
     var currentAmount: Int = 0,
-    var currentPrice: Int = 0
+    var currentPrice: Double = 0.0
 )
